Avoid signed overflow when assembling the 32-bit fields
data is uint8_t, so each element promotes to int before the shift, and a
byte of 0x80 or more shifted left by 24 overflows int.

Since C++20 that overflow is well defined and wraps, so the assembled
value is already correct where the addon is built as C++20. It pins no
standard -- CMakeLists.txt sets no CMAKE_CXX_STANDARD -- so it takes the
toolchain default, which is gnu++17 on gcc 14 and older, and there the
shift is undefined.

Cast the top byte to uint32_t so the shift happens in an unsigned type
and the result does not depend on which standard the toolchain picks.
The value produced is unchanged. The low three bytes cannot overflow and
are left alone.

+++ src/FluidDecoder.cpp
@@ -98,8 +98,10 @@ bool CFluidCodec::ReadTag(const std::string& filename, kodi::addon::AudioDecoder
 
   file.Read(data, len);
 
-  uint32_t header = data[3] | data[2] << 8 | data[1] << 16 | data[0] << 24;
-  uint32_t headerLength = data[7] | data[6] << 8 | data[5] << 16 | data[4] << 24;
+  uint32_t header = data[3] | data[2] << 8 | data[1] << 16 |
+                    static_cast<uint32_t>(data[0]) << 24;
+  uint32_t headerLength = data[7] | data[6] << 8 | data[5] << 16 |
+                          static_cast<uint32_t>(data[4]) << 24;
   if (header != MIDI_HEADER || headerLength != 6)
     return false;
 
@@ -111,10 +113,10 @@ bool CFluidCodec::ReadTag(const std::string& filename, kodi::addon::AudioDecoder
   std::string title;
   while (ptr < len)
   {
-    uint32_t trackHeader =
-        data[ptr + 3] | data[ptr + 2] << 8 | data[ptr + 1] << 16 | data[ptr] << 24;
-    uint32_t trackHeaderLength =
-        data[ptr + 7] | data[ptr + 6] << 8 | data[ptr + 5] << 16 | data[ptr + 4] << 24;
+    uint32_t trackHeader = data[ptr + 3] | data[ptr + 2] << 8 | data[ptr + 1] << 16 |
+                           static_cast<uint32_t>(data[ptr]) << 24;
+    uint32_t trackHeaderLength = data[ptr + 7] | data[ptr + 6] << 8 | data[ptr + 5] << 16 |
+                                 static_cast<uint32_t>(data[ptr + 4]) << 24;
 
     if (trackHeader != MIDI_MTrk)
       break;
